+++ include/prototypes.h
@@ -708,7 +708,7 @@ extern int prod_plane(int, int, struct bp *, int);
 extern void populace(struct natstr *, struct sctstr *, int);
 extern int total_work(int, int, int, int, int, int);
 /* prepare.c */
-extern void tax(struct sctstr *, struct natstr *, int, long *, int *,
+extern void tax(struct sctstr *, struct natstr *, int, int *, int *,
 		int *, int *);
 extern int upd_slmilcosts(natid, int);
 extern void prepare_sects(int, struct bp *);
